+++ Turbo2 Game Studio/Library/Graphics/Vector3D.h
@@ -3,29 +3,29 @@
 
 struct Vector3D
 {
-	float X, Y, Z;
+	double X, Y, Z;
 
 	//  Constructors
 	Vector3D();
-	Vector3D(float x, float y, float z);
+	Vector3D(double x, double y, double z);
 
-	float Length();
+	double Length();
 	Vector3D Normalize();
 
 	Vector3D operator-() const;
 
 	Vector3D operator+(const Vector3D &v) const;
     Vector3D operator-(const Vector3D &v) const;
-    Vector3D operator*(float r) const;
-    Vector3D operator/(float r) const;
+    Vector3D operator*(double r) const;
+    Vector3D operator/(double r) const;
 
-	Vector3D &operator=(float r);
+	Vector3D &operator=(double r);
 	Vector3D &operator+=(const Vector3D &v);
 	Vector3D &operator-=(const Vector3D &v);
-    Vector3D &operator*=(float r);
-    Vector3D &operator/=(float r);
+    Vector3D &operator*=(double r);
+    Vector3D &operator/=(double r);
 
-    float operator*(const Vector3D &v) const;   //  Dot product
+    double operator*(const Vector3D &v) const;   //  Dot product
 	Vector3D operator%(const Vector3D &v) const;  //  cross product
 };
