+++ Turbo2 Game Studio/Library/Scene/ITurboSceneObjectPlacement.h
@@ -5,7 +5,7 @@
 class ITurboSceneObjectPlacement
 {
 public:
-	//  ITurboSceneObjectPlacement Properties
+	//  ITurboSceneObjectPlacement Properties  ----------------------------------------------------------------------------
 	virtual Vector3D	Right() = 0;
 	virtual void		Right(Vector3D value) = 0;
 	virtual Vector3D	Up() = 0;
@@ -21,13 +21,13 @@ public:
 	virtual Vector3D	AngularVelocity() = 0;
 	virtual void		AngularVelocity(Vector3D value) = 0;
 
-	//  ITurboSceneObjectPlacement Methods
+	//  ITurboSceneObjectPlacement Methods  -------------------------------------------------------------------------------
 	virtual void	Reset() = 0;
-	virtual void	GoTo(float x, float y, float z) = 0;
+	virtual void	GoTo(double x, double y, double z) = 0;
 	virtual void	GoTo(Vector3D position) = 0;
-	virtual void	Move(float x, float y, float z) = 0;
+	virtual void	Move(double x, double y, double z) = 0;
 	virtual void	Move(Vector3D velocity) = 0;
-	virtual void	Rotate(float x, float y, float z) = 0;
+	virtual void	Rotate(double x, double y, double z) = 0;
 	virtual void	Rotate(Vector3D angularVelocity) = 0;
 	virtual std::shared_ptr<ITurboSceneObjectPlacement> Clone() = 0;
 };
